requirements.txt
flet>=0.25.0
openai-whisper
pydub
torch
torchaudio
pillow
pyinstaller
